default canopy param; test defaults
